refactoring prometheus addon
