+++ kogito-travel-agency/basic/src/main/java/org/kie/kogito/app/ProcessEventListenerConfig.java
@@ -15,7 +15,7 @@
  */
 package org.kie.kogito.app;
 
-import org.kie.kogito.monitoring.process.PrometheusProcessEventListener;
+import org.kie.kogito.monitoring.prometheus.process.PrometheusProcessEventListener;
 import org.kie.kogito.process.impl.DefaultProcessEventListenerConfig;
