+++ kogito-travel-agency/extended/travels/src/main/java/org/kie/kogito/app/RuleEventListenerConfig.java
@@ -18,7 +18,7 @@ package org.kie.kogito.app;
 import javax.enterprise.context.ApplicationScoped;
 
 import org.drools.core.config.DefaultRuleEventListenerConfig;
-import org.kie.kogito.monitoring.rule.PrometheusMetricsDroolsListener;
+import org.kie.kogito.monitoring.prometheus.rule.PrometheusMetricsDroolsListener;
 
 @ApplicationScoped
 public class RuleEventListenerConfig extends DefaultRuleEventListenerConfig {
